Check platform requirements
Instead of just showing the platform requirements, the platform
dependencies are now checked.

In order to assume the latest version of Composer being installed, which
has the corresponding new command available, a composer self-update was
added to the Composer validation step.

When all requirements are met, exit code 0 is emitted. If versions
mismatch, exit code 1 is emitted and when requirements are missing,
exit code 2 is emitted.