Feature Made composer validate strict
The reason for this is that the default pipeline also has this check built-in.
This way the testing-suite will give a more similar result to the pipelines.
I've also changed the composer.json file of the project, because it had @stable references in the requirements, this is not allowed anymore with the new setting.