Feature use wildcards for the namespaces for factory methods and the testframework
Within different projects there can be a dependency to different factory methods within other vendor namespaces.

With this commit the vendor part of the namespaces are made up of a wildcard.
This way it does not have to be explicitly defined. In normal use-cases it wouldn't be an issue to do so, because the factories are automatically generated.

http://devdocs.magento.com/guides/v2.0/extension-dev-guide/factories.html#writing-factories