Disable pdepend file cache
By using the memory cache implementation of pdepend it no longer stores
cache in ~/.pdepend. The file cache implementation has a garbage
collector for the cache which is executed every time pdepend is run. The
default TTL of the file cache is 30 days which results in a very large
cache directory. Scanning this cache directory for expired files in most
cases takes longer than using no cache.