[AD-210] Run phpcs with the -s flag to output the explicit rule that is failing.
When, for some reason, a sniff needs to be ignored it can be quite a hassle to figure out which sniff to ignore.
This commit adds the exact sniff code to the report.

E.g: a normal report would look like:

```
----------------------------------------------------------------------
FOUND 0 ERRORS AND 1 WARNING AFFECTING 1 LINE
----------------------------------------------------------------------
 29 | WARNING | Line exceeds 120 characters; contains 170 characters
----------------------------------------------------------------------
```

With this -s flag it will look like:

```
--------------------------------------------------------------------------------------------------------------
FOUND 0 ERRORS AND 1 WARNING AFFECTING 1 LINE
--------------------------------------------------------------------------------------------------------------
 29 | WARNING | Line exceeds 120 characters; contains 170 characters (Generic.Files.LineLength.TooLong)
--------------------------------------------------------------------------------------------------------------
```